Support for non-owning array views
The underlying problem with non-owning array views is that they can't be
used to allocate a new data buffer via one of their constructors.
However, ArrayHandlerIterable unconditionally required the presence of
such a constructor in its copyAny specialization for anything that is
neither std::array nor a plain C array.

Instead, we now provide a customization point in form of a function
template molpro::linalg::array::allocate_array which is used to allocate
new buffers for such arrays (views, really). The default implementation
simply throws an exception, meaning that it is only suitable for cases
where no new buffers have to be allocated. This is the case if a view is
used as a R container.

If needed, users can specialize this function for their view type in
order to actually perform the allocation in whatever way suitable for
their case.

Fixes #539